GNUmakefile: fix regression with default rule
It had been a while since I last ran "make" with no arguments, and it
was no longer triggering the "default" rule. This led to the observation
that user-supplied dependencies, such as the "ad-hoc" rule illustrated
in the GNUmakefile, need to occur after the call to do_mariner(). This
change fixes that and adds some explanation to the comments.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>