mariner: fix corner case in "delete" rules
We make deletion of a container image dependent on deletion of all
derived container images. (If you want to remove the 5th floor, you need
to first remove the 6th floor. Rinse and repeat.)

However that dependency should _not_ get generated if the set of derived
images is empty. This commit fixes that corner case.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>