examples: support corporate proxies, etc
Solves at least one known customization use-case, namely when docker build
and docker run need proxy settings and new trust roots (CA certificates)
to reach the internet.

We create a do-corporate.mk file (and add a small hook in GNUmakefile to
load it) to show how customization for "special environments" can occur.
This is intended as a first point of call for users that may require more
substantial tweaking, in the hope they might be able to avoid substantial
hacking of anything else.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>