mariner_v2.mk: tweak the IMAGE create/delete logic
If rebuilding an image that has already been successfully built, there is
some dependency reason for it, so we can't trust Docker to assume that
identical (Dockerfile) instructions mean we're in an identical situation.
(Environment, external proxying/caching realities, the Mariner build
system itself, ...)

Also, use --force-rm for "docker build", for similar rationale.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>