mariner_v2.mk: support asynchronous commands
Two new kinds of "RUNARGS", "detach_nojoin" and "detach_join", are
supported. The "nojoin" variety provides fire-and-forget behavior - the
container is detached and we do not keep track of it, it self-removes on
exit. The "join" variety generates two makefile targets, one to "launch"
the command in a container and detach (after recording the container ID
to a "joinfile"), and another to "wait" for the launched container to exit
(if it hasn't already). In fact, two targets are generated using
touchfiles that are produced and removed as a byproduct of oscillating
between the to-lauch/to-wait states. Optionally, two further targets can
also be produced that are named to be user-visible (and are simply
dependent on the touchfile-based, less-visible targets). This is a UX
consideration. (When desired, the named targets show up in bash's
tab-completion support for make. When not-desired, these named targets
can be omitted, such as when hundreds of async targets for some complex
test-case are auto-generated. You may only want targets visible for the
overall test-case, rather than each of its sub-elements.)

Also, fix volume dependencies. Only _managed_ volumes have create/delete
targets, so only declare dependencies on them for volumes that are
managed.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>