mariner_v2.mk: add support for NETWORKS
Rudimentary support for creation of networks (NETWORKS holds the global
list, like VOLUMES/IMAGES/etc), with attributes;
- DESCRIPTION (unused, like other DESCRIPTION attributes so far)
- DNAME (what to call the network inside docker, defaults to its name in
  Mariner)
- MANAGED (whether to create/delete in Mariner, otherwise externally-
  managed)
- XTRA (extra cmd-line options to pass to "docker run")

A container image will be connected to any/all networks listed in the
image's _NETWORKS attribute. This gets inherited into descendant images,
but this can be controlled by _UNNETWORKS. (The semantics are the same as
for associating volumes with images.)

Also, add a _HOSTNAME attribute to images, which defaults to the name of
the Mariner image object.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>