mariner_v2.mk: set --network-alias
User-created docker networks provide the appearance of DHCP and DNS. But,
by default, the container IDs are used for DNS purposes and these are
unpredictable hex strings in many modes of usage. We were already setting
the hostname (the name the container uses for itself), which the user can
specify, so use the same name to set a network-alias (the name served by
DNS to peer containers).

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>