features/user/as_me: fix sudo
If we depend on environment variables being injected for some reason, then
use of sudo requires the "-E" flag to avoid clobbering that environment.

NB: I have been a complete idiot in how I handle the "embed_me" user act
vs the root account in the containers, and will fix it shortly, which will
make the sudo usage disappear. But for the sake of posterity, I'm applying
the fix before applying the removal.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>