features/app/uml: teach UML to shutdown
Running user-mode-linux in a docker-like fashion takes a little effort. A
docker container runs its main process as PID 1, and considers the
container "done" when that process exits. A Linux system (e.g. that might
be able to run docker) runs a global oversight process called "init" as
PID 1, which has special handling and obligations (parenting orphaned
processes, etc) and which _does not exit_. Halting the system requires a
special system call (which also assumes coordinated cleanup of other
processes has already occurred).

This change adds the source code for a program that simply makes that
system call, and automates building and installing it to
/install/bin/myshutdown. If UML is invoked "docker-like", to run a single
command and then exit, a wrapper script should be invoked that runs the
required command and then invokes the myshutdown tool. That wrapper can
be provided to the UML instance with the "init=<path>" boot parameter.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>