features/debian/embed_me: improve group handling
This previously only worked in simple cases where the user and their
primary group (usually of the same name) were all that mattered. This
broke some cases, notably the features/mount/docker_socket layer, where
the non-root user needs to be a member of the "docker" group in order to
use the forwarded docker socket. This patch causes the account creation
script of "embed_me" to try to add the user (within the container) to the
same groups they belong to on the host side. Obviously, the container may
not have the same groups so this needs to ignore failures. A nasty
implication is the order with which layers are added. E.g. if you
"embed_me" _before_ adding the "docker_socket" layer, the "docker" group
won't exist and the attempt to add the user to the docker group will
fail. Then, when you add the "docker_socket" layer, the embedded user
can't (by default) use it.

Proofs of concepts get a free pass.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>