features/sys/null_passwd: new feature layer
Using Mariner+Docker to build root filesystems for use as more than just
containers is a powerful idea, but it does raise an issue that Docker
itself doesn't have, namely that starting up the system doesn't inherently
"log you in". (Nor does it inherently "run the thing you wanted to run",
nor does it inherently "shutdown when the thing you wanted to run is back
from its run", etc.) Starting a container inherently starts the command
that the container was spawned to run, the idea of "logging in" doesn't
occur, therefore passwords are largely a non-subject.

More explanation in the comments at the top of null_passwd.mk, but
basically this feature layer gives container images a passwd/shadow
configuration in which the root account has an empty password.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>