features/sys/nada: new "empty" feature layer
This layer does 'nada', literally. The idea is that you may have
conditional logic in a GNUmakefile that is building up a stack of images
in succession, and if the names are fixed but you want to "skip" one of
the steps only in a particular situation, how do you do so without
breaking the chain of image derivations (by name)? Well, you use the
"sys/nada" layer instead of whatever layer you're trying to skip. It
replaces the unwanted link in the chain, by producing an output/child
image that is unmodified from the input/parent image.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>