examples/user-mode-linux: fixes and improvements
When creating the 'rootfs' Docker image, for conversion into a filesystem
image for use by UML, we want to incorporate the "myshutdown" utility that
was built by the "app/uml" layer (in the 'uml' Docker image) and installed
into the /install volume. This change adds the dependencies to ensure the
'uml' image is created, its 'build' verb has been run, and the resulting
'myshutdown' binary is put into the Docker context for the 'rootfs' image.

Also, use 'doobie' instead of 'debbie' as the wrapper for debian:latest.
Generated makefile rules that deal with "touch" files can otherwise seem
inappropriate, which wasn't and isn't my intention.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>