examples/alcatraz: jail-style usage
Probably the most "traditional usage of Docker" example of Mariner, it
creates an image for apps that may need a little more containment than
usual. This creates a persistent home-directory mount to give it a place
to store dotfiles and what-not, but keeps that distinct from the home
account on the host. Then, assuming these apps should to be able to read
files from your host account's ~/Downloads and ~/Documents directories,
uses unmanaged, read-only mounts of those into spawned containers. Finally,
a hacky "run-arbitrary" scheme is added, allowing (almost) arbitrary
commands to be started up in the container without creating new
verbs/commands and (host shims) for them individually.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>