mariner_v2.mk: DSPACE == "Docker namespace"
DSPACE defines a prefix to use for Docker objects, behind the scenes. I.e.
IMAGES and NETWORKS will continue to be named in Makefiles and in
user-exposed Makefile targets "as-is", but entities that get created in
and manipulated by Docker will have those names prefixed by "$(DSPACE)_",
as a way of modularising and insulating use-cases from each other. If the
caller defines DSPACE, that's what it will be. Otherwise, it defaults to
$(shell basename `pwd`), i.e. the name of the current directory.

Signed-off-by: Geoff Thorpe <geoff@geoffthorpe.net>